implement op functions

+++ install.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 if [ -z "$1" ]; then
     echo "installing mpm"
-    curl -s https://raw.githubusercontent.com/paperbenni/mpm/master/mpm.sh | sudo tee /usr/local/bin/mpm
+    if ! [ -e ./mpm.sh ]; then
+        curl -s https://raw.githubusercontent.com/paperbenni/mpm/master/mpm.sh | sudo tee /usr/local/bin/mpm
+    else
+        cat mpm.sh | sudo tee /usr/local/bin/mpm
+    fi
     sudo chmod 755 /usr/local/bin/mpm
     echo "done installing mpm"
 else
